requirements.txt
numpy
biopython
argparse
